homebrew optimizations + execution modify

+++ src/commands/brew/backup.rs
@@ -6,7 +6,10 @@ use clap::Args;
 use toml_edit::{Array, DocumentMut, Item, Table, Value};
 
 use crate::{
-    brew::utils::{brew_list, disable_auto_update, ensure_brew, restore_auto_update},
+    brew::{
+        types::BrewListType,
+        utils::{brew_list, disable_auto_update, ensure_brew, restore_auto_update},
+    },
     commands::Runnable,
     config::get_config_path,
     util::{
@@ -26,29 +29,21 @@ pub struct BrewBackupCmd {
 impl Runnable for BrewBackupCmd {
     async fn run(&self) -> Result<()> {
         let cfg_path = get_config_path();
+        let dry_run = should_dry_run();
+        let mut prev = None;
 
         // disable auto-update
-        let prev = disable_auto_update();
+        if !dry_run {
+            prev = disable_auto_update();
+        }
 
         // ensure brew install
         ensure_brew().await?;
 
-        let formulas = brew_list(&["list", "--formula"]).await?;
-        let casks = brew_list(&["list", "--cask"]).await?;
-        let deps = brew_list(&["list", "--installed-as-dependency"]).await?;
-        let taps = brew_list(&["tap"]).await?;
-
-        if should_dry_run() {
-            print_log(
-                LogLevel::Dry,
-                &format!(
-                    "Would backup {} formulae and {} casks",
-                    formulas.len(),
-                    casks.len()
-                ),
-            );
-            return Ok(());
-        }
+        let formulas = brew_list(BrewListType::Formula).await?;
+        let casks = brew_list(BrewListType::Cask).await?;
+        let taps = brew_list(BrewListType::Tap).await?;
+        let mut deps = Vec::new();
 
         let mut doc = if cfg_path.exists() {
             let text = fs::read_to_string(&cfg_path).await?;
@@ -64,6 +59,7 @@ impl Runnable for BrewBackupCmd {
         // firstly remember the --no-deps value
         brew_tbl["no-deps"] = Item::None;
         if self.no_deps {
+            deps = brew_list(BrewListType::Dependency).await?;
             print_log(
                 LogLevel::Info,
                 "Setting no-deps to true in config for later reads.",
@@ -76,15 +72,26 @@ impl Runnable for BrewBackupCmd {
         for formula in &formulas {
             if self.no_deps {
                 if !deps.contains(formula) {
-                    print_log(
-                        LogLevel::Info,
-                        &format!("Pushing {} as a manually installed formula.", formula),
-                    );
-                    formula_arr.push(formula.as_str());
+                    if dry_run {
+                        print_log(
+                            LogLevel::Dry,
+                            &format!("Would push {} as a manually installed formula.", formula),
+                        );
+                    } else {
+                        print_log(
+                            LogLevel::Info,
+                            &format!("Pushing {} as a manually installed formula.", formula),
+                        );
+                        formula_arr.push(formula.as_str());
+                    }
                 }
             } else {
-                print_log(LogLevel::Info, &format!("Pushing {}", formula));
-                formula_arr.push(formula.as_str());
+                if dry_run {
+                    print_log(LogLevel::Dry, &format!("Would push {}", formula));
+                } else {
+                    print_log(LogLevel::Info, &format!("Pushing {}", formula));
+                    formula_arr.push(formula.as_str());
+                }
             }
         }
         print_log(
@@ -95,8 +102,29 @@ impl Runnable for BrewBackupCmd {
 
         let mut cask_arr = Array::new();
         for cask in &casks {
-            print_log(LogLevel::Info, &format!("Pushed {} as a cask.", cask));
-            cask_arr.push(cask.as_str());
+            if self.no_deps {
+                if !deps.contains(cask) {
+                    if dry_run {
+                        print_log(
+                            LogLevel::Dry,
+                            &format!("Would push {} as a manually installed cask.", cask),
+                        );
+                    } else {
+                        print_log(
+                            LogLevel::Info,
+                            &format!("Pushing {} as a manually installed cask.", cask),
+                        );
+                        cask_arr.push(cask.as_str());
+                    }
+                }
+            } else {
+                if dry_run {
+                    print_log(LogLevel::Dry, &format!("Would push {}", cask));
+                } else {
+                    print_log(LogLevel::Info, &format!("Pushed {} as a cask.", cask));
+                    cask_arr.push(cask.as_str());
+                }
+            }
         }
         print_log(LogLevel::Info, &format!("Pushed {} casks.", cask_arr.len()));
         brew_tbl["casks"] = Item::Value(Value::Array(cask_arr));
@@ -104,27 +132,35 @@ impl Runnable for BrewBackupCmd {
         // backup taps
         let mut taps_arr = Array::new();
         for tap in &taps {
-            print_log(LogLevel::Info, &format!("Pushed {} as a tap.", tap));
-            taps_arr.push(tap.as_str());
+            if dry_run {
+                print_log(LogLevel::Dry, &format!("Would push {} as tap.", tap));
+            } else {
+                print_log(LogLevel::Info, &format!("Pushed {} as a tap.", tap));
+                taps_arr.push(tap.as_str());
+            }
         }
         print_log(LogLevel::Info, &format!("Pushed {} taps.", taps_arr.len()));
         brew_tbl["taps"] = Item::Value(Value::Array(taps_arr));
 
-        // give length of both lists in verbose, and let the user know about config location
-        print_log(LogLevel::Info, &format!("Writing backup to {:?}", cfg_path));
-        fs::write(&cfg_path, doc.to_string()).await?;
+        // write backup
+        if !dry_run {
+            fs::write(&cfg_path, doc.to_string()).await?;
+            print_log(LogLevel::Info, &format!("Backup saved to {:?}", cfg_path));
+            print_log(
+                LogLevel::Fruitful,
+                &format!(
+                    "Done! You can find the backup in your config file location {:?}",
+                    cfg_path
+                ),
+            );
+            restore_auto_update(prev);
+        } else {
+            print_log(
+                LogLevel::Info,
+                &format!("Backup would be saved to {:?}", cfg_path),
+            );
+        }
 
-        // output message
-        print_log(LogLevel::Info, &format!("Backup saved to {:?}", cfg_path));
-        print_log(
-            LogLevel::Fruitful,
-            &format!(
-                "Done! You can find the backup in your config file location {:?}",
-                cfg_path
-            ),
-        );
-
-        restore_auto_update(prev);
         Ok(())
     }
 }
